some code cleaning + added coverage
